[#76] Enhancement: Add user login
When you are a user (as in a human being) using the cli, you should always
be yourself. The existing login mechanism only allows for running as a
machine with the use of client id and secrets. While this is a valid (and
is still the default) way of using this cli, a user login type has been
added.

You can now add a login profile with:
mim login add --type=user --server=<DataHub> --authorizer="https://auth.mimiro.io"

This will then prompt you to log in via your browser, and once you do, it
will use the refresh token provided to get a valid token, instead of id/secret.

To enable this, a lot of code had to be moved around, and some new code
was also added. To avoid a large rewrite, old code was wrapped around the old
where appropriate. There are now new ways of doing for example web calls,
but old ones still work fine.

The config system was rewritten to wrap it in a lib, as it is now used
quite a bit more places.