[FEATURE] Allow job operate on multiple job ids
When you want to enable or disable or do any other operation on jobs that
involves more than 1 job, you have a lot of extra work to do today.
This change allows to input multiple ids as input to job operate.

Example:
```bash
mim jobs operate -o run \
    --id proxy-cima-production-to-local \
    --id renewing-slayback \
    --id sacred-lionheart
```

The only exception is if you run a job with -w (follow) set, then you can
only give 1 id. This is because of the ui differences between this and
the other operations. These differences can be worked out for sure, but is
not in scope for this change.

The api.JobManager has been extended with an Operate service that contains
the methods for job operations.

The web client has been extended with web.Get[T any]() and web.Put[T any]()
to make it easier to parse result from http calls into structs. I only added
these, as the web client(s) can really benefit from a full refactor to get
it under control.

I also added a "renderkit" file under utils, this is intended to keep functions
used to render the ui, for now it just contains a utils.Date() func that
formats a date to local timezone.